requirements.txt
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-nested-routers==0.93.3
flake8==3.9.2
flake8-html==0.4.1
